+++ HISTORY.md
@@ -1,5 +1,19 @@
 # History
 
+## 0.1.2 (2019-05-07)
+
+REST API and Cluster Management.
+
+### New Features
+
+* REST API Server - Issues [#82](https://github.com/HDI-Project/ATM/issues/82) and
+  [#132](https://github.com/HDI-Project/ATM/issues/132) by @RogerTangos, @pvk-developer and @csala
+* Add Cluster Management commands to start and stop the server and multiple workers
+  as background processes - Issue [#130](https://github.com/HDI-Project/ATM/issues/130) by
+  @pvk-developer and @csala
+* Add TravisCI and migrate docs to GitHub Pages - Issue [#129](https://github.com/HDI-Project/ATM/issues/129)
+  by @pvk-developer
+
 ## 0.1.1 (2019-04-02)
 
 First Release on PyPi.